requirements.txt
astroid==2.4.2
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pip-autoremove==0.9.1
psycopg2==2.8.5
pylint==2.5.3
pymongo==3.10.1
python-dotenv==0.14.0
six==1.15.0
toml==0.10.1
wrapt==1.12.1
